requirements.txt
flask==2.0.1
requests==2.26.0
fluent-logger
kafka-python==2.0.2
networkx==2.6.3
opentelemetry-api==1.12.0
opentelemetry-sdk==1.12.0
opentelemetry-instrumentation-flask==0.33b0
opentelemetry-instrumentation-requests==0.33b0
elasticsearch==8.10.0
python-json-logger==2.0.7
colorama